refactor : vo 패키지를 domain 내부로 이동
