+++ src/main/java/racingcar/domain/vo/Trial.java
@@ -1,4 +1,4 @@
-package racingcar.vo;
+package racingcar.domain.vo;
 
 import java.util.Objects;
