+++ src/test/java/racingcar/domain/CarTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.vo.CarName;
+import racingcar.domain.vo.CarName;
 
 import static org.assertj.core.api.Assertions.assertThat;
