+++ src/test/java/racingcar/vo/CarNameTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
+import racingcar.domain.vo.CarName;
 
 import static org.assertj.core.api.Assertions.*;
