+++ src/test/java/racingcar/vo/PositionTest.java
@@ -3,9 +3,9 @@ package racingcar.vo;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.vo.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.in;
 
 class PositionTest {
     @Nested
